Styled the price check overlay
