+++ src/Sidekick/Windows/Overlay/UserControls/ItemListingControl.xaml
@@ -1,32 +1,10 @@
-<UserControl x:Class="Sidekick.Windows.Overlay.UserControls.ItemListingControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             mc:Ignorable="d">
-    <d:DesignerProperties.DesignStyle>
-        <Style TargetType="UserControl">
-            <Setter Property="Background"  Value="Gray" />
-            <Setter Property="Width"  Value="480" />
-            <Setter Property="Height"  Value="Auto" />
-        </Style>
-    </d:DesignerProperties.DesignStyle>
-
-    <UserControl.Resources>
-        <SolidColorBrush x:Key="TextColor" Color="#FFE6E6E6" />
-        
-        <Style TargetType="{x:Type TextBox}">
-            <Setter Property="BorderBrush" Value="Transparent" />
-            <Setter Property="BorderThickness" Value="0" />
-            <Setter Property="Background" Value="Transparent" />
-            <Setter Property="Foreground" Value="{DynamicResource TextColor}" />
-            <Setter Property="IsReadOnly" Value="True" />
-            <Setter Property="TextWrapping" Value="NoWrap" />
-            <Setter Property="Padding" Value="5,0" />
-            <Setter Property="Margin" Value="0 4" />
-        </Style>
-    </UserControl.Resources>
-
+<UserControl
+    x:Class="Sidekick.Windows.Overlay.UserControls.ItemListingControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="140*" />
@@ -35,11 +13,26 @@
             <ColumnDefinition Width="40*" />
             <ColumnDefinition Width="80*" />
         </Grid.ColumnDefinitions>
-            
-        <TextBox Grid.Column="0" Name="accountNameTextbox" Text="AccountName" />
-        <TextBox Grid.Column="1" Name="characterNameTextbox" Text="CharacterName" />
-        <TextBox Grid.Column="2" Name="priceTextbox" Text="Price" VerticalAlignment="Center" />
-        <TextBox Grid.Column="3" Name="itemLevelTextbox" Text="iLvl" HorizontalAlignment="Right" />
-        <TextBox Grid.Column="4" Name="ageTextbox" Text="Age" HorizontalAlignment="Right" />
+
+        <TextBlock
+            Name="accountNameTextbox"
+            Grid.Column="0"
+            Text="AccountName" />
+        <TextBlock
+            Name="characterNameTextbox"
+            Grid.Column="1"
+            Text="CharacterName" />
+        <TextBlock
+            Name="priceTextbox"
+            Grid.Column="2"
+            Text="Price" />
+        <TextBlock
+            Name="itemLevelTextbox"
+            Grid.Column="3"
+            Text="iLvl" />
+        <TextBlock
+            Name="ageTextbox"
+            Grid.Column="4"
+            Text="Age" />
     </Grid>
 </UserControl>
