--- src/Sidekick/Windows/Overlay/UserControls/QueryResultControl.xaml
@@ -1,99 +1,0 @@
-<UserControl
-    x:Class="Sidekick.Windows.Overlay.UserControls.QueryResultControl"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-    <d:DesignerProperties.DesignStyle>
-        <Style TargetType="UserControl">
-            <Setter Property="Background" Value="Gray" />
-            <Setter Property="Width" Value="480" />
-            <Setter Property="Height" Value="Auto" />
-            <Setter Property="Label.Content" Value="White" />
-            <Setter Property="Label.Foreground" Value="Red" />
-        </Style>
-    </d:DesignerProperties.DesignStyle>
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition />
-            <RowDefinition />
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition />
-            <ColumnDefinition />
-        </Grid.ColumnDefinitions>
-
-        <TextBox
-            Grid.Row="0"
-            Margin="5,5,0,0"
-            FontWeight="Bold"
-            Foreground="{Binding Item.Rarity, Converter={StaticResource RarityToColorConverter}}"
-            Text="{Binding Path=Item.Name, Mode=OneWay}" />
-
-        <TextBox
-            Grid.Row="0"
-            Grid.Column="1"
-            Margin="5,0,5,0"
-            FontWeight="Bold"
-            TextAlignment="Right">
-            <TextBox.Style>
-                <Style BasedOn="{StaticResource {x:Type TextBox}}" TargetType="{x:Type TextBox}">
-                    <Setter Property="Foreground" Value="White" />
-                    <Setter Property="Visibility" Value="Hidden" />
-                    <Style.Triggers>
-                        <DataTrigger Binding="{Binding HasAverage}" Value="True">
-                            <Setter Property="Visibility" Value="Visible" />
-                        </DataTrigger>
-                    </Style.Triggers>
-                </Style>
-            </TextBox.Style>
-            <TextBox.Text>
-                <MultiBinding StringFormat="poe.ninja: {0}c">
-                    <Binding Path="PoeNinjaItem.ChaosValue" />
-                </MultiBinding>
-            </TextBox.Text>
-        </TextBox>
-        <TextBlock
-            Grid.Row="1"
-            Grid.Column="1"
-            Margin="8,0,5,0"
-            TextAlignment="Right">
-            <TextBlock.Style>
-                <Style BasedOn="{StaticResource {x:Type TextBlock}}" TargetType="{x:Type TextBlock}">
-                    <Setter Property="Visibility" Value="Hidden" />
-                    <Style.Triggers>
-                        <DataTrigger Binding="{Binding HasAverage}" Value="True">
-                            <Setter Property="Visibility" Value="Visible" />
-                        </DataTrigger>
-                    </Style.Triggers>
-                </Style>
-            </TextBlock.Style>
-            <TextBlock.Text>
-                <MultiBinding StringFormat="Last update: {0:HH:mm}">
-                    <Binding Path="LastRefreshTimestamp" />
-                </MultiBinding>
-            </TextBlock.Text>
-        </TextBlock>
-
-        <TextBlock Grid.Row="1" Grid.Column="0">
-            <Hyperlink
-                NavigateUri="{Binding Uri}"
-                RequestNavigate="openQueryLink"
-                TextDecorations="None">
-                <TextBlock
-                    Grid.Row="1"
-                    Margin="8,0,0,5"
-                    TextDecorations="Underline">
-                    <TextBlock.Text>
-                        <MultiBinding StringFormat="{}Showing {0} items out of {1}">
-                            <Binding Mode="OneWay" Path="Result.Count" />
-                            <Binding Path="Total" />
-                        </MultiBinding>
-                    </TextBlock.Text>
-                </TextBlock>
-            </Hyperlink>
-        </TextBlock>
-    </Grid>
-</UserControl>
