--- src/Sidekick/Windows/Overlay/UserControls/QueryResultControl.xaml.cs
@@ -1,22 +1,0 @@
-using System.Diagnostics;
-using System.Windows.Controls;
-using System.Windows.Navigation;
-
-namespace Sidekick.Windows.Overlay.UserControls
-{
-    public partial class QueryResultControl : UserControl
-    {
-        public QueryResultControl()
-        {
-            InitializeComponent();
-        }
-
-        private void openQueryLink(object sender, RequestNavigateEventArgs e)
-        {
-            var uri = e.Uri.ToString();
-            Legacy.Logger.Log(string.Format("Opening in browser: {0}", uri));
-            Process.Start(uri);
-            e.Handled = true;
-        }
-    }
-}
